pyproject.toml
[project]
name = "bet-analysis"
version = "0.1.0"
description = "Pre-match betting analysis MCP server with multi-AI debate (Claude + GPT + Gemini)"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "ereencaan" }]
dependencies = [
    "mcp>=1.2.0",
    "httpx>=0.27.0",
    "pydantic>=2.7.0",
    "aiosqlite>=0.20.0",
    "python-dotenv>=1.0.1",
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "google-genai>=0.3.0",
]

[project.scripts]
bet-analysis = "bet_analysis.server:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "vcrpy>=6.0.0",
    "ruff>=0.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bet_analysis"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
